Create hotels controller and setup google maps
A hotels controller is created with an index action which will be
used to show a list of results when a user searchs in the home page
a google map is implemented to allow show results locations in the
future. Some customizations like a navbar are implemented in a custom
layout. The index action route of the hotel's controller is aliased
as s to be similiar to airbnb